feat: Legit Profiler more refactored, stop profilling works
