+++ src/engine/editor/LegitProfiler/ProfilerGraph.hpp
@@ -4,11 +4,11 @@
 
 #pragma once
 
-#include "ProfilerTask.h"
+#include "legitProfiler.hpp"
+#include "ProfilerTask.hpp"
 #include "imgui.h"
 #include <algorithm>
 #include <array>
-#include <chrono>
 #include <glm/fwd.hpp>
 #include <glm/gtx/string_cast.hpp>
 #include <ios>
@@ -17,9 +17,28 @@
 #include <sstream>
 #include <vector>
 
-namespace ImGuiUtils {
-inline glm::vec2 Vec2(ImVec2 vec) { return glm::vec<2, float>(vec.x, vec.y); }
+
+namespace legit {
+
 class ProfilerGraph {
+  private:
+  struct FrameData {
+    std::vector<legit::ProfilerTask> tasks;
+    std::vector<size_t> taskStatsIndex;
+  };
+
+  struct TaskStats {
+    double maxTime;
+    size_t priorityOrder;
+    size_t onScreenIndex;
+  };
+
+  std::vector<TaskStats> taskStats;
+  std::map<std::string, size_t> taskNameToStatsIndex;
+
+  std::vector<FrameData> frames;
+  size_t currFrameIndex = 0;
+
   public:
   int frameWidth;
   int frameSpacing;
@@ -74,7 +93,7 @@ class ProfilerGraph {
 
   void RenderTimings(int graphWidth, int legendWidth, int height, int frameIndexOffset) {
     ImDrawList *drawList = ImGui::GetWindowDrawList();
-    const glm::vec2 widgetPos = Vec2(ImGui::GetCursorScreenPos());
+    const glm::vec2 widgetPos = to_glm_vec(ImGui::GetCursorScreenPos());
     RenderGraph(drawList, widgetPos, glm::vec2(graphWidth, height), frameIndexOffset);
     RenderLegend(drawList, widgetPos + glm::vec2(graphWidth, 0.0f), glm::vec2(legendWidth, height),
                  frameIndexOffset);
@@ -111,6 +130,7 @@ class ProfilerGraph {
       taskStats[statIndex].priorityOrder = statNumber;
     }
   }
+
   void RenderGraph(ImDrawList *drawList, glm::vec2 graphPos, glm::vec2 graphSize,
                    size_t frameIndexOffset) {
     Rect(drawList, graphPos, graphPos + graphSize, 0xffffffff, false);
@@ -128,8 +148,8 @@ class ProfilerGraph {
       glm::vec2 taskPos = framePos + glm::vec2(0.0f, 0.0f);
       auto &frame = frames[frameIndex];
       for (const auto &task : frame.tasks) {
-        float taskStartHeight = task.startTime * graphSize.y / (1./30 * 1000);
-        float taskEndHeight = task.endTime * graphSize.y / (1./30 * 1000);
+        float taskStartHeight = task.startTime * graphSize.y / (maxFrameTime * 1000);
+        float taskEndHeight = task.endTime * graphSize.y / (maxFrameTime * 1000);
 
         if (std::abs(taskEndHeight - taskStartHeight) > heightThreshold) {
           Rect(drawList, taskPos + glm::vec2(0.0f, -taskStartHeight),
@@ -138,6 +158,7 @@ class ProfilerGraph {
       }
     }
   }
+
   void RenderLegend(ImDrawList *drawList, glm::vec2 legendPos, glm::vec2 legendSize,
                     size_t frameIndexOffset) {
     float markerLeftRectMargin = 3.0f;
@@ -171,8 +192,8 @@ class ProfilerGraph {
         stat.onScreenIndex = tasksShownCount++;
       } else
         continue;
-      float taskStartHeight = task.startTime * legendSize.y / (1./30 * 1000);
-      float taskEndHeight = task.endTime * legendSize.y / (1./30 * 1000);
+      float taskStartHeight = task.startTime * legendSize.y / (maxFrameTime * 1000);
+      float taskEndHeight = task.endTime * legendSize.y / (maxFrameTime * 1000);
 
       glm::vec2 markerLeftRectMin = legendPos + glm::vec2(markerLeftRectMargin, legendSize.y);
       glm::vec2 markerLeftRectMax = markerLeftRectMin + glm::vec2(markerLeftRectWidth, 0.0f);
@@ -184,7 +205,6 @@ class ProfilerGraph {
           glm::vec2(markerLeftRectMargin + markerLeftRectWidth + markerMidWidth,
                     legendSize.y - markerRigthRectMargin -
                         (markerRightRectHeight + markerRightRectSpacing) * stat.onScreenIndex
-                    //20
                     );
       glm::vec2 markerRightRectMax =
           markerRightRectMin + glm::vec2(markerRightRectWidth, -markerRightRectHeight);
@@ -192,7 +212,7 @@ class ProfilerGraph {
                        markerRightRectMax, task.color);
 
       uint32_t textColor =
-          useColoredLegendText ? task.color : legit::Colors::imguiText; // task.color;
+          useColoredLegendText ? task.color : legit::Colors::imguiText;
 
       float taskTimeMs = float(task.endTime - task.startTime);
       std::ostringstream timeText;
@@ -203,72 +223,6 @@ class ProfilerGraph {
       Text(drawList, markerRightRectMax + textMargin + glm::vec2(nameOffset, 0.0f), textColor,
            (std::string("ms] ") + task.name).c_str());
     }
-
-    /*
-    struct PriorityEntry
-    {
-      bool isUsed;
-      legit::ProfilerTask task;
-    };
-    std::map<std::string, PriorityEntry> priorityEntries;
-    for (auto priorityTask : priorityTasks)
-    {
-      PriorityEntry entry;
-      entry.task = frames[priorityTask.frameIndex].tasks[priorityTask.taskIndex];
-      entry.isUsed = false;
-      priorityEntries[entry.task.name] = entry;
-    }
-    size_t shownTasksCount = 0;
-    for (size_t taskIndex = 0; taskIndex < currFrame.tasks.size(); taskIndex++)
-    {
-      auto &task = currFrame.tasks[taskIndex];
-      auto it = priorityEntries.find(task.name);
-      if (it != priorityEntries.end() && !it->second.isUsed)
-      {
-        it->second.isUsed = true;
-
-        float taskStartHeight = (float(task.startTime) / maxFrameTime) * legendSize.y;
-        float taskEndHeight = (float(task.endTime) / maxFrameTime) * legendSize.y;
-
-
-        glm::vec2 markerLeftRectMin = legendPos + glm::vec2(markerLeftRectMargin, legendSize.y);
-        glm::vec2 markerLeftRectMax = markerLeftRectMin + glm::vec2(markerLeftRectWidth, 0.0f);
-        markerLeftRectMin.y -= taskStartHeight;
-        markerLeftRectMax.y -= taskEndHeight;
-
-        glm::vec2 markerRightRectMin = legendPos + glm::vec2(markerLeftRectMargin +
-    markerLeftRectWidth + markerMidWidth, legendSize.y - markerRigthRectMargin -
-    (markerRightRectHeight + markerRightRectSpacing) * shownTasksCount); glm::vec2
-    markerRightRectMax = markerRightRectMin + glm::vec2(markerRightRectWidth,
-    -markerRightRectHeight); RenderTaskMarker(drawList, markerLeftRectMin, markerLeftRectMax,
-    markerRightRectMin, markerRightRectMax, task.color);
-
-        uint32_t textColor = legit::Colors::imguiText;// task.color;
-
-        float taskTimeMs = float(task.endTime - task.startTime);
-        std::ostringstream timeText;
-        timeText.precision(2);
-        timeText << std::fixed << std::string("[") << (taskTimeMs * 1000.0f);
-
-        Text(drawList, markerRightRectMax + textMargin, textColor, timeText.str().c_str());
-        Text(drawList, markerRightRectMax + textMargin + glm::vec2(nameOffset, 0.0f), textColor,
-    (std::string("ms] ") + task.name).c_str()); shownTasksCount++;
-      }
-    }*/
-
-    /*for (size_t priorityTaskIndex = 0; priorityTaskIndex < priorityTasks.size();
-    priorityTaskIndex++)
-    {
-      auto &priorityTask = priorityTasks[priorityTaskIndex];
-      auto &globalTask = frames[priorityTask.frameIndex].tasks[priorityTask.taskIndex];
-
-      size_t lastFrameTaskIndex = currFrame.FindTask(globalTask.name);
-
-      glm::vec2 taskPos = legendPos + marginSpacing + glm::vec2(0.0f, markerHeight) +
-    glm::vec2(0.0f, (markerHeight + itemSpacing) * priorityTaskIndex); Rect(drawList, taskPos,
-    taskPos + glm::vec2(markerHeight, -markerHeight), task.color, true); Text(drawList, taskPos +
-    textOffset, 0xffffffff, task.name.c_str());
-    }*/
   }
 
   static void Rect(ImDrawList *drawList, glm::vec2 minPoint, glm::vec2 maxPoint, uint32_t col,
@@ -278,9 +232,11 @@ class ProfilerGraph {
     else
       drawList->AddRect(ImVec2(minPoint.x, minPoint.y), ImVec2(maxPoint.x, maxPoint.y), col);
   }
+
   static void Text(ImDrawList *drawList, glm::vec2 point, uint32_t col, const char *text) {
     drawList->AddText(ImVec2(point.x, point.y), col, text);
   }
+
   static void Triangle(ImDrawList *drawList, std::array<glm::vec2, 3> points, uint32_t col,
                        bool filled = true) {
     if (filled)
@@ -291,6 +247,7 @@ class ProfilerGraph {
       drawList->AddTriangle(ImVec2(points[0].x, points[0].y), ImVec2(points[1].x, points[1].y),
                             ImVec2(points[2].x, points[2].y), col);
   }
+
   static void RenderTaskMarker(ImDrawList *drawList, glm::vec2 leftMinPoint, glm::vec2 leftMaxPoint,
                                glm::vec2 rightMinPoint, glm::vec2 rightMaxPoint, uint32_t col) {
     Rect(drawList, leftMinPoint, leftMaxPoint, col, true);
@@ -300,124 +257,5 @@ class ProfilerGraph {
         ImVec2(rightMinPoint.x, rightMaxPoint.y), ImVec2(rightMinPoint.x, rightMinPoint.y)};
     drawList->AddConvexPolyFilled(points.data(), int(points.size()), col);
   }
-  struct FrameData {
-    /*void BuildPriorityTasks(size_t maxPriorityTasksCount)
-    {
-      priorityTaskIndices.clear();
-      std::set<std::string> usedTaskNames;
-
-      for (size_t priorityIndex = 0; priorityIndex < maxPriorityTasksCount; priorityIndex++)
-      {
-        size_t bestTaskIndex = size_t(-1);
-        for (size_t taskIndex = 0; taskIndex < tasks.size(); taskIndex++)
-        {
-          auto &task = tasks[taskIndex];
-          auto it = usedTaskNames.find(tasks[taskIndex].name);
-          if (it == usedTaskNames.end() && (bestTaskIndex == size_t(-1) ||
-    tasks[bestTaskIndex].GetLength() < task.GetLength()))
-          {
-            bestTaskIndex = taskIndex;
-          }
-        }
-        if (bestTaskIndex == size_t(-1))
-          break;
-        priorityTaskIndices.push_back(bestTaskIndex);
-        usedTaskNames.insert(tasks[bestTaskIndex].name);
-      }
-    }*/
-    std::vector<legit::ProfilerTask> tasks;
-    std::vector<size_t> taskStatsIndex;
-    // std::vector<size_t> priorityTaskIndices;
-  };
-
-  struct TaskStats {
-    double maxTime;
-    size_t priorityOrder;
-    size_t onScreenIndex;
-  };
-  std::vector<TaskStats> taskStats;
-  std::map<std::string, size_t> taskNameToStatsIndex;
-
-  /*struct PriorityTask
-  {
-    size_t frameIndex;
-    size_t taskIndex;
-  };
-  std::vector<PriorityTask> priorityTasks;*/
-  std::vector<FrameData> frames;
-  size_t currFrameIndex = 0;
 };
-
-class ProfilersWindow {
-  public:
-  ProfilersWindow() : cpuGraph(300), gpuGraph(300) {
-    stopProfiling = false;
-    frameOffset = 0;
-    frameWidth = 3;
-    frameSpacing = 1;
-    useColoredLegendText = true;
-    prevFpsFrameTime = std::chrono::system_clock::now();
-    fpsFramesCount = 0;
-    avgFrameTime = 1.0f;
-  }
-
-  void Render() {
-    fpsFramesCount++;
-    auto currFrameTime = std::chrono::system_clock::now();
-    {
-      float fpsDeltaTime = std::chrono::duration<float>(currFrameTime - prevFpsFrameTime).count();
-      if (fpsDeltaTime > 0.5f) {
-        this->avgFrameTime = fpsDeltaTime / float(fpsFramesCount);
-        fpsFramesCount = 0;
-        prevFpsFrameTime = currFrameTime;
-      }
-    }
-
-    ImVec2 canvasSize = ImGui::GetContentRegionAvail();
-
-    int sizeMargin = int(ImGui::GetStyle().ItemSpacing.y);
-    int maxGraphHeight = 300;
-    int availableGraphHeight = (int(canvasSize.y) - sizeMargin) / 2;
-    int graphHeight = std::min(maxGraphHeight, availableGraphHeight);
-    int legendWidth = 200;
-    int graphWidth = int(canvasSize.x) - legendWidth;
-    gpuGraph.RenderTimings(graphWidth, legendWidth, graphHeight, frameOffset);
-    cpuGraph.RenderTimings(graphWidth, legendWidth, graphHeight, frameOffset);
-    if (graphHeight * 2 + sizeMargin + sizeMargin < canvasSize.y) {
-      ImGui::Columns(2);
-      ImGui::Checkbox("Stop profiling", &stopProfiling);
-      cpuGraph.stopProfiling = stopProfiling;
-      gpuGraph.stopProfiling = stopProfiling;
-      // ImGui::SetNextItemWidth(ImGui::GetContentRegionAvail().x - textSize);
-      ImGui::Checkbox("Colored legend text", &useColoredLegendText);
-      ImGui::DragInt("Frame offset", &frameOffset, 1.0f, 0, 400);
-      ImGui::NextColumn();
-
-      ImGui::SliderInt("Frame width", &frameWidth, 1, 4);
-      ImGui::SliderInt("Frame spacing", &frameSpacing, 0, 2);
-      ImGui::SliderFloat("Transparency", &ImGui::GetStyle().Colors[ImGuiCol_WindowBg].w, 0.0f,
-                         1.0f);
-      ImGui::Columns(1);
-    }
-    if (!stopProfiling)
-      frameOffset = 0;
-    gpuGraph.frameWidth = frameWidth;
-    gpuGraph.frameSpacing = frameSpacing;
-    gpuGraph.useColoredLegendText = useColoredLegendText;
-    cpuGraph.frameWidth = frameWidth;
-    cpuGraph.frameSpacing = frameSpacing;
-    cpuGraph.useColoredLegendText = useColoredLegendText;
-  }
-  bool stopProfiling;
-  int frameOffset;
-  ProfilerGraph cpuGraph;
-  ProfilerGraph gpuGraph;
-  int frameWidth;
-  int frameSpacing;
-  bool useColoredLegendText;
-  using TimePoint = std::chrono::time_point<std::chrono::system_clock>;
-  TimePoint prevFpsFrameTime;
-  size_t fpsFramesCount;
-  float avgFrameTime;
-};
-} // namespace ImGuiUtils
+}
