+++ src/engine/editor/ProfilingWindow.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "LegitProfiler/ImGuiProfilerRenderer.h"
+#include "ProfilerWindow.hpp"
 #include "glWindow.hpp"
 #include "EditorWindow.hpp"
 #include <chrono>
@@ -20,11 +20,12 @@ namespace cevy::editor {
 class ProfilingWindow : public EditorWindow {
     std::vector<float> frames;
     std::vector<float> framesAverage;
-    ImGuiUtils::ProfilersWindow legitProfiler;
+    legit::ProfilerWindow legitProfiler;
     std::chrono::high_resolution_clock::time_point last_call;
+
     public:
     ProfilingWindow() : EditorWindow(true, "Profiling") {
-      legitProfiler = ImGuiUtils::ProfilersWindow();
+      legitProfiler = legit::ProfilerWindow();
     }
 
     void render(cevy::editor::Editor &editor, glWindow &glwindow) override;
